pyproject.toml
[project]
name = "aggregator-domain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg-pool>=3.2.6",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.0",
    "tldextract>=5.3.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.10",
]
